% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retell_text.R
\name{retell_text}
\alias{retell_text}
\title{Title}
\usage{
retell_text(url = NULL, token = NULL)
}
\arguments{
\item{url}{URL you want to retell}

\item{token}{Your Token}
}
\value{
List
}
\description{
Title
}
\examples{
retell_text(url = 'your_url', token = 'your_token')
}
